feat: add link to tx for order notifications
We don’t always have the txId in the order information, but when we do
we add a link to the notification.
It’s highly likely that order created events won’t ever have a txId as
this event is created before submitting the tx to the blockchain.

I had to extract getTxLink to a separate file as having it in
useTxNotifications was causing a circular dependency error in our tests.
