+++ src/hooks/useTxNotifications.ts
@@ -1,14 +1,12 @@
 import { useEffect, useMemo, useRef } from 'react'
 import { formatError } from '@/utils/formatters'
-import type { LinkProps } from 'next/link'
 import { selectNotifications, showNotification } from '@/store/notificationsSlice'
 import { useAppDispatch, useAppSelector } from '@/store'
 import { TxEvent, txSubscribe } from '@/services/tx/txEvents'
-import { AppRoutes } from '@/config/routes'
 import { useCurrentChain } from './useChains'
 import useTxQueue from './useTxQueue'
 import { isSignableBy, isTransactionListItem } from '@/utils/transaction-guards'
-import { type ChainInfo, TransactionStatus } from '@safe-global/safe-gateway-typescript-sdk'
+import { TransactionStatus } from '@safe-global/safe-gateway-typescript-sdk'
 import { selectPendingTxs } from '@/store/pendingTxsSlice'
 import useIsSafeOwner from '@/hooks/useIsSafeOwner'
 import useWallet from './wallets/useWallet'
@@ -16,6 +14,7 @@ import useSafeAddress from './useSafeAddress'
 import { getExplorerLink } from '@/utils/gateway'
 import { getTxDetails } from '@/services/transactions'
 import { isWalletRejection } from '@/utils/wallets'
+import { getTxLink } from '@/utils/tx-link'
 
 const TxNotifications = {
   [TxEvent.SIGN_FAILED]: 'Failed to sign. Please try again.',
@@ -42,20 +41,6 @@ enum Variant {
 }
 
 const successEvents = [TxEvent.PROPOSED, TxEvent.SIGNATURE_PROPOSED, TxEvent.ONCHAIN_SIGNATURE_SUCCESS, TxEvent.SUCCESS]
-
-export const getTxLink = (
-  txId: string,
-  chain: ChainInfo,
-  safeAddress: string,
-): { href: LinkProps['href']; title: string } => {
-  return {
-    href: {
-      pathname: AppRoutes.transactions.tx,
-      query: { id: txId, safe: `${chain?.shortName}:${safeAddress}` },
-    },
-    title: 'View transaction',
-  }
-}
 
 const useTxNotifications = (): void => {
   const dispatch = useAppDispatch()
