+++ src/store/swapOrderSlice.ts
@@ -6,6 +6,8 @@ import { isSwapTxInfo, isTransactionListItem } from '@/utils/transaction-guards'
 import { txHistorySlice } from '@/store/txHistorySlice'
 import { showNotification } from '@/store/notificationsSlice'
 import { selectSafeInfo } from '@/store/safeInfoSlice'
+import { selectChainById } from '@/store/chainsSlice'
+import { getTxLink } from '@/utils/tx-link'
 
 type AllStatuses = OrderStatuses | 'created'
 type Order = {
@@ -80,11 +82,18 @@ export const swapOrderStatusListener = (listenerMiddleware: typeof listenerMiddl
       if (oldStatus === newStatus || newStatus === undefined) {
         return
       }
+      const safeInfo = selectSafeInfo(listenerApi.getState())
+
+      let link = null
+      if (swapOrder.txId && safeInfo.data?.chainId && safeInfo.data?.address) {
+        const chainInfo = selectChainById(listenerApi.getState(), safeInfo.data?.chainId)
+        if (chainInfo !== undefined) {
+          link = getTxLink(swapOrder.txId, chainInfo, safeInfo.data?.address.value)
+        }
+      }
 
       switch (newStatus) {
         case 'created':
-          const safeInfo = selectSafeInfo(listenerApi.getState())
-
           dispatch(
             showNotification({
               title: 'Order created',
@@ -94,6 +103,7 @@ export const swapOrderStatusListener = (listenerMiddleware: typeof listenerMiddl
                   : 'Waiting for confirmation from signers of your Safe',
               groupKey,
               variant: 'info',
+              link: link ? link : undefined,
             }),
           )
 
@@ -105,6 +115,7 @@ export const swapOrderStatusListener = (listenerMiddleware: typeof listenerMiddl
               message: 'Waiting for confirmation from signers of your Safe',
               groupKey,
               variant: 'info',
+              link: link ? link : undefined,
             }),
           )
           break
@@ -115,6 +126,7 @@ export const swapOrderStatusListener = (listenerMiddleware: typeof listenerMiddl
               message: 'Waiting for order execution by the CoW Protocol',
               groupKey,
               variant: 'info',
+              link: link ? link : undefined,
             }),
           )
           break
@@ -129,6 +141,7 @@ export const swapOrderStatusListener = (listenerMiddleware: typeof listenerMiddl
               message: 'Your order has been successful',
               groupKey,
               variant: 'success',
+              link: link ? link : undefined,
             }),
           )
           break
@@ -143,6 +156,7 @@ export const swapOrderStatusListener = (listenerMiddleware: typeof listenerMiddl
               message: 'Your order has reached the expiry time and has become invalid',
               groupKey,
               variant: 'warning',
+              link: link ? link : undefined,
             }),
           )
           break
@@ -157,6 +171,7 @@ export const swapOrderStatusListener = (listenerMiddleware: typeof listenerMiddl
               message: 'Your order has been cancelled',
               groupKey,
               variant: 'warning',
+              link: link ? link : undefined,
             }),
           )
           break
